backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
google-genai>=0.2.0
google-cloud-aiplatform>=1.60.0
google-auth>=2.23.0
python-multipart>=0.0.9
orjson>=3.10.7
python-dotenv>=1.0.1
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
opencv-python>=4.10.0
numpy>=1.26.0

